config: remove legacy app config resolver
